Create readability.c
Readability problem involves creating a program in C to assess the readability of text. It calculates the approximate grade level required to comprehend a given passage. By counting letters, words, and sentences, the program applies the Coleman-Liau index to estimate the reading level. This exercise sharpens skills in string manipulation, character counting, and algorithmic computation